Refactor PropPaneHelp to use npmFunctions presets
