+++ src/webparts/drilldown7/CoreFPS/PreConfiguredSettings.ts
